pyproject.toml
[project]
name = "tenantapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.120.0",
    "orjson>=3.11.4",
    "pydantic[email]>=2.12.3",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.3",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "uvicorn>=0.38.0",
]
[project.scripts]
dev = "src.scripts.server:run_server"
start = "src.scripts.server:run_prod"
seed = "src.scripts.seed:seed"


[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src",

]
